% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_2_numbers}
\alias{add_2_numbers}
\title{Add two numbers, create new class}
\usage{
add_2_numbers(x, y, ...)
}
\arguments{
\item{x}{the first number}

\item{y}{the second number}

\item{...}{Additional arguments to pass through}
}
\value{
A list of class \code{my_new_class}
}
\description{
This is an example of how to create a new class.
}
