% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{boot.selector.ic}
\alias{boot.selector.ic}
\title{Inference for selector_ic class based on bootstrap
other methods are supposed in infer.selector.ic}
\usage{
boot.selector.ic(
  model,
  B = 10,
  nonselection = "ignored",
  parallel = FALSE,
  make_levels = FALSE,
  ...
)
}
\arguments{
\item{model}{model returned from selector_ic  class}

\item{B}{Number of bootstraps samples}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{parallel}{Whether to run bootstrap in parallel or not}

\item{...}{Additional arguments}
}
\value{
"boot_ic" class list with
\item{model}{A dataframe with model results including columns for term, mean_estimate,  conf.low, conf.high, median_p.value,ci_ln, prop.select, prop.reject}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{selected the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
\item{B}{The number of bootstrap replicates used}
}
\description{
Inference for selector_ic class based on bootstrap
other methods are supposed in infer.selector.ic
}
