% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.selector_pen}
\alias{boot.selector_pen}
\title{Bootstrapping for penalized CV-based model}
\usage{
\method{boot}{selector_pen}(
  model,
  B = 250,
  nonselection = "ignored",
  parallel = FALSE,
  boot_desparse = FALSE,
  save_beta = FALSE,
  ...
)
}
\arguments{
\item{model}{model of class selector_pen}

\item{B}{The number of bootstrap replicates.}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{parallel}{should a cluster with detectCores()-1 be made and used}

\item{boot_desparse}{TBD}

\item{...}{any additional arguments to that can be passed to ncvreg}

\item{family}{Currently only "gaussian" supported}
}
\value{
A tidy dataframe with bootstrap dataset
\item{term}{variable name}
\item{mean_estimate}{mean of regression coefficients across bootstrap samples}
\item{conf.low}{lower 2.5 percentile bootstrap interval}
\item{conf.high}{upper 97.5 percentile bootstrap interval}
\item{median_p.value}{median p value of regression coefficients  across bootstrap samples}
\item{ci_ln}{confidence interval length}
\item{prop.select}{proportion of times a given variable is selected by model selection method}
\item{prop.rej}{proportion of time coefficient was found significant at 0.05 alpha level}
}
\description{
Bootstrapping for penalized CV-based model
}
