% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_internal.R
\name{boot_pen_parallel}
\alias{boot_pen_parallel}
\title{Internal function for bootstrapping penalized model in parallel
Currently this could be made more efficient.}
\usage{
boot_pen_parallel(
  x,
  y,
  B,
  family,
  std,
  nonselection,
  n_cores,
  parallel,
  lambda_selected,
  nlambda,
  all_terms,
  lmax,
  penalty,
  alpha,
  selected_terms,
  lambda_seq,
  boot_desparse,
  save_beta
)
}
\description{
Internal function for bootstrapping penalized model in parallel
Currently this could be made more efficient.
}
