% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_internal.R
\name{boot_stepwise}
\alias{boot_stepwise}
\title{Internal function for bootstrapping}
\usage{
boot_stepwise(
  x,
  y,
  family,
  penalty,
  std,
  B,
  nonselector,
  direction,
  make_levels,
  save_beta,
  ...
)
}
\description{
Internal function for bootstrapping
}
