% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fns.R
\name{boot_stepwise_bic}
\alias{boot_stepwise_bic}
\title{Title}
\usage{
boot_stepwise_bic(
  x,
  y,
  B = 250,
  family = "gaussian",
  nonselector = "ignored",
  parallel = FALSE,
  direction = "both",
  model = model,
  ...
)
}
\arguments{
\item{x}{dataframe with predictors}

\item{y}{outcome vector}

\item{B}{The number of bootstrap replicates.}

\item{family}{Currently only "gaussian" supported}

\item{nonselector}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{parallel}{whether to run bootstrap in parallel for faster computation}

\item{direction}{the mode of stepwise search, can be one of "both", "backward", or "forward",
with a default of "both". If the scope argument is missing the default for direction is "backward".}

\item{model}{Selected model from whole data set and stepwise BIC method}

\item{...}{any additional arguments to that can be passed to stepAIC}
}
\value{
Tidy dataframe with bootstrap result and CIs
\item{term}{variable name}
\item{mean_estimate}{mean of regression coefficients across bootstrap samples}
\item{conf.low}{lower 2.5 percentile bootstrap interval}
\item{conf.high}{upper 97.5 percentile bootstrap interval}
\item{median_p.value}{median p value of regression coefficients  across bootstrap samples}
\item{ci_ln}{confidence interval length}
\item{prop.select}{propotion of times a given variable is selected by model selection method}
\item{prop.rej}{proportion of time coefficient was found significant at 0.05 alpha level}
}
\description{
Title
}
