% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collat_fn.R
\name{ciratio}
\alias{ciratio}
\title{Computing mean/median CI ratio and significant Discoveries}
\usage{
ciratio(
  formula,
  data,
  family = c("gaussian", "binomial", "poisson"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls"),
  conf.level = 0.95,
  B = 250,
  direction = "forward",
  select_factors_together = F,
  debias = F,
  inference_target = "selections",
  ...
)
}
\arguments{
\item{formula}{a formula}

\item{data}{data set}

\item{family}{outcome distributional family}

\item{nonselection}{A character string specifying how to handle variables
not selected by model selection procedure. One of "ignored",
"confident_nulls" or "uncertain_nulls" supported}

\item{conf.level}{.95 by default}

\item{B}{Number of bootstraps}

\item{direction}{the mode of step wise search, can be one of "both", "backward", or "forward", with a default of "forward"}

\item{select_factors_together}{should categorical variables be jointly selected?}

\item{debias}{should estimates be debiased in bootstrap}

\item{inference_target}{is inference requested on all or selected only}

\item{...}{additional arguments}
}
\value{
A list with following data frames
\item{avg_ci_ln}{Mean of CI length across all variables in model excluding intercept}
\item{med_ci_ln}{Median of CI length across all variables in model excluding intercept}
\item{no_sign_disc}{# of significant discoveries in model excluding intercept}
}
\description{
Perfroms full model boostrap, stepwise AIC, stepwise BIc, lasso and MCP and gives average CI for variables
adjusting for post-selectino inference using hybrid, bootstrap and selective inference
}
