% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{clean_colnames_fn}
\alias{clean_colnames_fn}
\title{Title}
\usage{
clean_colnames_fn(nms)
}
\arguments{
\item{nms}{a vector with column names to be cleaned}
}
\value{
a vector with clean columns
}
\description{
Title
}
