% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_helpers.R
\name{dummy_col_fn}
\alias{dummy_col_fn}
\title{Title}
\usage{
dummy_col_fn(data, make_levels)
}
\arguments{
\item{data}{data to be processed}

\item{make_levels}{whether to make dummy variables for each factor level}
}
\value{
model matrix with clean dataset
}
\description{
Title
}
