% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fns.R
\name{full_boot}
\alias{full_boot}
\title{Title}
\usage{
full_boot(
  model,
  B = 250,
  family = "gaussian",
  parallel = FALSE,
  std = F,
  save_beta = F,
  ...
)
}
\arguments{
\item{model}{lm object with design matrix and output variable}

\item{B}{The number of bootstrap replicates.}

\item{family}{currently only "gaussian" supported}

\item{parallel}{whether to run bootstrap in parallel for faster computation}
}
\value{
A tidy dataframe with bootstrap result and CIs
\item{term}{variable name}
\item{mean_estimate}{mean of regression coefficients across bootstrap samples}
\item{conf.low}{lower 2.5 percentile bootstrap interval}
\item{conf.high}{upper 97.5 percentile bootstrap interval}
\item{median_p.value}{median p value of regression coefficients  across bootstrap samples}
\item{ci_ln}{confidence interval length}
\item{prop.rej}{proportion of time coefficient was found significant at 0.05 alpha level}
\item{ci_avg_ratio}{Average of CI length across all variables after bootstrap}
\item{ci_median_ratio}{median of CI length across all variables after bootstrap}
}
\description{
Title
}
