% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hers}
\alias{hers}
\title{HERS data example}
\format{
A data frame with \code{n} rows and \code{p} variables:
\describe{
\item{age}{age in years}
\item{bmi}{BMI (kg/m^2)}
\item{dbp}{ diastolic blood pressure.}
\item{diabetes}{diabetes}
\item{dmpills}{ oral DM medication by self-report}
\item{drinkany}{any current alcohol consumption}
\item{exercise}{exercise at least 3 times per week.}
\item{globrat}{self-reported health}
\item{glucose}{fasting glucose (mg/dl) }
\item{hdl}{  HDL cholesterol (mg/dl)}
\item{hdl1}{year 1 HDL cholesterol (mg/dl).}
\item{ht}{random assignment to hormone therapy}
\item{htnmeds}{ anti-hypertensive use}
\item{insulin}{insulin use by self-report}
\item{ldl}{LDL cholesterol (mg/dl)}
\item{medcond}{other serious conditions by self-report.}
\item{physact}{comparative physical activity.}
\item{raceth}{race/ethnicity.}
\item{sbp}{systolic blood pressure}
\item{smoking}{current smoker}
\item{statins}{statin use}
\item{tg}{triglycerides (mg/dl)}
\item{waist}{waist (cm).}
\item{weight}{weight (kg)}
\item{whr}{ waist/hip ratio.}
}
}
\source{
\describe{
\item{hers}{\url{insert-url-here}}
}
}
\usage{
hers
}
\description{
This dataset contains data from the HERS study used for demonstration purposes.
Missing data (7-8\%) has been omitted.
}
\examples{
data(hers)
head(hers)
}
\references{
\describe{
Add reference here
}
}
\keyword{datasets}
