% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.selector.ic}
\alias{infer.selector.ic}
\title{Inference for selector.ic class except bootstrap which is it's own function}
\usage{
infer.selector.ic(model, method = "hybrid", nonselection = "ignored")
}
\arguments{
\item{model}{model returned from selector_ic  class}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid"or  "selectiveinf"}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
infer_ic class list with
\item{model} {A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio} {Average CI length across all variables in model}
\item{ci_median_ratio} {medain CI length across all variables in model}
\item{selection_method} {Stepwsie}
\item{direction} {the mode of step wise search}
\item{penalty} {penalty used (AIC or BIC)}
\item{infmethod} {Inference method chosen}
\item{nonselection} {method chosen  to deal with non selection}
}
\description{
Inference for selector.ic class except bootstrap which is it's own function
}
