% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.selector.pen}
\alias{infer.selector.pen}
\title{Inference for selector.pen class except bootstrap which is it's own function}
\usage{
infer.selector.pen(model, method = "hybrid", nonselection = "ignored")
}
\arguments{
\item{model}{model of selector_pen class returned from  pen_cv function}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid", "selectiveinf" or
"boot" supported}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
"infer_pen" class list with
}
\description{
Inference for selector.pen class except bootstrap which is it's own function
}
