% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_inferrer.R
\name{tidy.inferrer}
\alias{tidy.inferrer}
\alias{print.inferrer}
\title{tidy method for \code{inferrer} object}
\usage{
\method{tidy}{inferrer}(x, scale_coef = TRUE, ...)

\method{print}{inferrer}(x, ...)
}
\arguments{
\item{x}{an \code{inferrer}}

\item{scale_coef}{should scaled betas be returned, or unscaled?}

\item{...}{additional parameters (not yet used)}
}
\description{
Returns a tibble with all candidate variables, estimates (scaled & unscaled)
}
