% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_cv.R
\name{pen_cv}
\alias{pen_cv}
\title{Title}
\usage{
pen_cv(x, y, std = TRUE, penalty = "lasso", lambda = "min", ...)
}
\arguments{
\item{x}{Dataframe with predictors}

\item{y}{outcome vector}

\item{std}{if TRUE (defult), standardize design matrix}

\item{penalty}{lasso or MCP}

\item{lambda}{extrac coefficeints assoicated wtih lambda min or lamba SE}

\item{...}{Additional arguments that can be passed}
}
\value{
A data frame with betahat
}
\description{
Title
}
