% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{print.boot_ic}
\alias{print.boot_ic}
\title{Title}
\usage{
\method{print}{boot_ic}(x, ...)
}
\arguments{
\item{x}{model of class \code{boot_ic}}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Title
}
