% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{print.infer_boot_selector_ic}
\alias{print.infer_boot_selector_ic}
\title{Title}
\usage{
\method{print}{infer_boot_selector_ic}(x, ...)

\method{print}{infer_boot_selector_ic}(x, ...)
}
\arguments{
\item{x}{x model of class \code{infer_boot_selector_pen}}

\item{...}{currently not used}
}
\value{
A tibble containing the tidied coefficients of the model.

returns x invisibly
}
\description{
Title

Title
}
