% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{print.infer_pen}
\alias{print.infer_pen}
\title{Title}
\usage{
\method{print}{infer_pen}(x, ...)
}
\arguments{
\item{x}{x model of class \code{infer_pen}}
}
\value{
returns x invisibly
}
\description{
Title
}
