% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_ic.R
\name{print.infer_stepwise_ic}
\alias{print.infer_stepwise_ic}
\title{Title}
\usage{
\method{print}{infer_stepwise_ic}(x, ...)
}
\arguments{
\item{x}{model of class \code{infer_stepwise_ic}}
}
\value{
returns x invisibly
}
\description{
Title
}
