% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ic.R
\name{print.selector_ic}
\alias{print.selector_ic}
\title{Title}
\usage{
\method{print}{`selector_ic`}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_ic}}

\item{...}{additional arguments (currently not supported)}
}
\value{
returns x invisibly
}
\description{
Title
}
