% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_cv.R
\name{print.selector_pen}
\alias{print.selector_pen}
\title{Title}
\usage{
\method{print}{selector_pen}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_pen}}
}
\value{
returns x invisibly
}
\description{
Title
}
