% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_pval.R
\name{print.selector_two_step_pval}
\alias{print.selector_two_step_pval}
\title{Title}
\usage{
\method{print}{`selector_two_step_pval`}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_two_step_pval}}

\item{...}{additional arguments (currently not supported)}
}
\value{
returns x invisibly
}
\description{
Title
}
