% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sel_inf}
\alias{sel_inf}
\title{Selective inference for lasso (internal helper)}
\usage{
sel_inf(x, y, model, lam = "lambda.min", intercept = TRUE, alpha = 1, ...)
}
\arguments{
\item{x}{design matrix}

\item{y}{outcome variable}

\item{lam}{lambda.1se or lambda.min}

\item{...}{addtional arguments that can be passed to  cv.glmnet}

\item{std}{whether to standaddize design matrix}
}
\value{
returns a data frame with results from lasso selective inference
}
\description{
Selective inference for lasso (internal helper)
}
