% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_full_model.R
\name{select_full_model}
\alias{select_full_model}
\title{Select a full model}
\usage{
select_full_model(
  formula,
  data,
  family = c("gaussian", "binomial", "poisson"),
  fitted_selector = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula.}

\item{data}{data set}

\item{family}{outcome distributional family}

\item{fitted_selector}{a previously fit \code{selector}, used for resampling}

\item{...}{Additional arguments}
}
\value{
A  \code{selector} object
}
\description{
This function simply attempts a full model and returns a \code{selector} object
as a useful comparison.
}
