% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ncvreg.R
\name{select_ncvreg}
\alias{select_ncvreg}
\title{Fit an MCP- or SCAD-penalized regression via \code{ncvreg}}
\usage{
select_ncvreg(
  formula,
  data,
  family = c("gaussian", "binomial", "poisson"),
  lambda = c("best", "compact"),
  fitted_selector = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula}

\item{data}{data set}

\item{family}{outcome distributional family}

\item{lambda}{can be \code{best}, \code{compact} (which use CV), or a numeric vector. See details.}

\item{fitted_selector}{a previously fit \code{selector}, used for resampling}

\item{...}{Additional arguments that can be passed to \code{ncvreg}, e.g. \code{penalty}, \code{alpha}}
}
\value{
A \code{selector} object wrapping \code{ncvreg}
}
\description{
A wrapper for \code{ncvreg} and \code{cv.ncvreg}. It runs CV by default so remember to
set your seed for reproducibility.
}
