% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_data}
\alias{std_data}
\title{Standadrized data Example}
\format{
A data frame with \code{n} rows and \code{p} variables:
\describe{
\item{age}{age in years}
\item{bmi}{BMI (kg/m^2)}
\item{dbp}{ diastolic blood pressure.}
\item{diabetes}{diabetes}
\item{dmpills}{ oral DM medication by self-report}
\item{drinkany}{any current alcohol consumption}
\item{exercise}{exercise at least 3 times per week.}
\item{globrat}{self-reported health}
\item{glucose}{fasting glucose (mg/dl) }
\item{hdl}{  HDL cholesterol (mg/dl)}
\item{hdl1}{year 1 HDL cholesterol (mg/dl).}
\item{ht}{random assignment to hormone therapy}
\item{htnmeds}{ anti-hypertensive use}
\item{insulin}{insulin use by self-report}
\item{ldl}{LDL cholesterol (mg/dl)}
\item{medcond}{other serious conditions by self-report.}
\item{physact}{comparative physical activity.}
\item{raceth}{race/ethnicity.}
\item{sbp}{systolic blood pressure}
\item{smoking}{current smoker}
\item{statins}{statin use}
\item{tg}{triglycerides (mg/dl)}
\item{waist}{waist (cm).}
\item{weight}{weight (kg)}
\item{whr}{ waist/hip ratio.}
}
}
\source{
\describe{
\item{std_data}{\url{}}
}
}
\usage{
std_data
}
\description{
This dataset contains raw data where continous varibles are standadrized and it is  used for demonstration purposes.
}
\examples{
data(std_data)
head(std_data)
}
\references{
\describe{
}
}
\keyword{datasets}
