% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ic.R
\name{step_ic}
\alias{step_ic}
\title{Stepwise forwad/backward/bidirectional selection with AIC/BIC}
\usage{
step_ic(x, y, std = TRUE, penalty = "AIC", direction = "both", ...)
}
\arguments{
\item{x}{Dataframe with predictors}

\item{y}{outcome vector}

\item{std}{if TRUE (defult), standardize design matrix}

\item{penalty}{AIC or BIC}

\item{direction}{the mode of stepwise search, can be one of "both", "backward", or "forward", with a default of "both"}

\item{...}{Additional arguments that can be passed}
}
\value{
A data frame with betahat
}
\description{
#' @description
}
\details{
This function implements forwad/backward/bidirectional stepwise regression,
for use in the practicalPSI package

You can learn more about package authoring with RStudio at:
}
