% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{tidy.boot_ic}
\alias{tidy.boot_ic}
\title{Title}
\usage{
\method{tidy}{boot_ic}(x, ...)
}
\arguments{
\item{x}{model of class \code{boot_ic}}

\item{...}{currently not used}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Title
}
