% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{tidy.infer_boot_selector_pen}
\alias{tidy.infer_boot_selector_pen}
\title{Tidy for bootstrapped post-regularized inference objects}
\usage{
\method{tidy}{infer_boot_selector_pen}(x, ...)
}
\arguments{
\item{x}{model of class \code{infer_boot_selector_pen}}

\item{...}{currently not used}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Tidy for bootstrapped post-regularized inference objects
}
