% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{tidy.infer_pen}
\alias{tidy.infer_pen}
\title{Title}
\usage{
\method{tidy}{infer_pen}(x, ...)
}
\arguments{
\item{x}{model of class \code{infer_pen}}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Title
}
