% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ic.R
\name{tidy.selector_ic}
\alias{tidy.selector_ic}
\title{Title}
\usage{
\method{tidy}{selector_ic}(x, ...)
}
\arguments{
\item{x}{A \code{selector_ic} object to tidy.}

\item{...}{Additional arguments passed to the generic \code{tidy} function.}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Title
}
