% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_cv.R
\name{tidy.selector_pen}
\alias{tidy.selector_pen}
\title{Title}
\usage{
\method{tidy}{selector_pen}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_pen}}

\item{...}{currently not used}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Title
}
