% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_ic.R
\name{tidy.selector_stepwise_ic}
\alias{tidy.selector_stepwise_ic}
\title{Tidy-er for selector object (\code{selector_stepwise_ic})}
\usage{
\method{tidy}{selector_stepwise_ic}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_stepwise_ic}}

\item{...}{Additional arguments passed to the generic \code{tidy} function.}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Tidy-er for selector object (\code{selector_stepwise_ic})
}
