% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_pval.R
\name{tidy.selector_two_step_pval}
\alias{tidy.selector_two_step_pval}
\title{Tidy-er for selector object (\code{selector_two_step_pval})}
\usage{
\method{tidy}{selector_two_step_pval}(x, ...)
}
\arguments{
\item{x}{model of class \code{selector_two_step_pval}}

\item{...}{Additional arguments passed to the generic \code{tidy} function.}
}
\value{
A tibble containing the tidied coefficients of the model.
}
\description{
Tidy-er for selector object (\code{selector_two_step_pval})
}
