% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_step_pval.R
\name{two_step_pval}
\alias{two_step_pval}
\title{A two-step selection process using p-values}
\usage{
two_step_pval(
  x,
  y,
  family = "gaussian",
  std = FALSE,
  sig.level = 0.05,
  starting_mod = c("null", "full"),
  ...
)
}
\arguments{
\item{x}{Dataframe/model matrix with predictors (without intercept)}

\item{y}{outcome vector}

\item{family}{currently gaussian supported}

\item{std}{if TRUE (default), standardize design matrix}

\item{sig.level}{significance level for inclusion (default: 0.05)}

\item{starting_mod}{the mode of the search, can be "null" or "full" for forward vs backward}

\item{...}{Additional arguments (not used currently)}
}
\value{
A list of class \code{selector_ic} containing:

\item{beta}{a tibble containing term names and coefficients}
\item{std}{Was desing matrix standardized}
\item{sig.level}{penalty used (significance level for inclusion)}
\item{starting_mod}{the mode of the search ("null" or "full")}
\item{x}{the model dataframe used}
\item{y}{repsonse used in vector}
\item{model_sum}{the selected model details}
}
\description{
This function implements a two-step selection process, either
screening marginal associations (null to model) or filtering (full to model).
}
