% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_boot.R
\name{infer_boot}
\alias{infer_boot}
\alias{boot}
\title{Inference for \code{selector} via bootstrapping}
\usage{
infer_boot(
  object,
  data,
  inference_target = c("selections", "all"),
  debias = FALSE,
  estimation_data = c("in-sample", "out-of-sample"),
  conf.level = 0.95,
  type = c("paired", "residual"),
  B = 250,
  n_cores = 4,
  ...
)

boot(
  object,
  data,
  B,
  inference_target = c("selections", "all"),
  debias = FALSE,
  estimation_data = c("in-sample", "out-of-sample"),
  conf.level,
  n_cores,
  ...
)
}
\arguments{
\item{object}{a \code{selector} object}

\item{data}{data must be passed to infer}

\item{inference_target}{is inference requested on all or selected only}

\item{debias}{should estimates be debiased (no, non-selections, or all)}

\item{estimation_data}{within a bootstrap, should in-sample or
out-of-sample residuals be used for estimation (defaults to FALSE)}

\item{conf.level}{.95 by default}

\item{type}{what type of bootstrap (currently only \code{paired} supported)}

\item{B}{The number of bootstrap replicates.}

\item{n_cores}{number of cores for parallel computation}

\item{...}{any additional arguments to that can be passed to fitting engine}
}
\value{
\code{inferrer} s3 class with things like...

an \code{inferrer} object
}
\description{
Inference for \code{selector} via bootstrapping

Bootstrapping selection process
}
