% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot}
\alias{boot}
\alias{boot.selector_ic}
\title{Bootstrapping model selection methods (case-wise)}
\usage{
boot(
  model,
  B = 250,
  nonselection = c("ignored", "uncertain_nulls", "confident_nulls"),
  n_cores = 1,
  boot_desparse = FALSE,
  make_levels = FALSE,
  save_beta = FALSE,
  ...
)

\method{boot}{selector_ic}(
  model,
  B = 250,
  nonselection = "ignored",
  n_cores = 1,
  make_levels = FALSE,
  save_beta = FALSE,
  ...
)
}
\arguments{
\item{model}{Selected model from whole data set and stepwise AIC method}

\item{B}{The number of bootstrap replicates.}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{n_cores}{whether to run bootstrap in parallel for faster computation}

\item{boot_desparse}{for penalized methods, should within-bootstrap coefficients be refit?}

\item{make_levels}{?}

\item{save_beta}{?}

\item{...}{any additional arguments to that can be passed to stepAIC}

\item{nonselector}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
Ddataframe with bootstrap result and CIs
\item{term}{variable name}
\item{mean_estimate}{mean of regression coefficients across bootstrap samples}
\item{conf.low}{lower 2.5 percentile bootstrap interval}
\item{conf.high}{upper 97.5 percentile bootstrap interval}
\item{median_p.value}{median p value of regression coefficients  across bootstrap samples}
\item{ci_ln}{confidence interval length}
\item{prop.select}{propotion of times a given variable is selected by model selection method}
\item{prop.rej}{proportion of time coefficient was found significant at 0.05 alpha level}
}
\description{
Bootstrapping model selection methods (case-wise)

Bootstrapping for selection process with stepwise AIC/BIC
}
