% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_boot.R
\name{infer_boot}
\alias{infer_boot}
\alias{boot}
\title{Inference for \code{selector} via bootstrapping}
\usage{
infer_boot(
  object,
  data,
  type = c("paired", "residual"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls"),
  conf.level = 0.95,
  B = 250,
  n_cores = 4,
  ...
)

boot(
  object,
  data,
  B,
  nonselection,
  uncertain_insample = FALSE,
  conf.level,
  n_cores,
  ...
)
}
\arguments{
\item{object}{a \code{selector} object}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{conf.level}{.95 by default}

\item{B}{The number of bootstrap replicates.}

\item{n_cores}{number of cores for parallel computation}

\item{...}{any additional arguments to that can be passed to fitting engine}

\item{uncertain_insample}{within a bootstrap, should in-sample or
out-of-sample residuals be used for estimation (defaults to FALSE)}

\item{model}{model returned from selector_stepwise_ic  class}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid"or  "selectiveinf"}
}
\value{
infer s3 class with things like...
\item{selector}{the selector used to fit the model}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{medain CI length across all variables in model}
\item{infmethod}{Inference method}
\item{nonselection}{method chosen  to deal with non selection}

Ddataframe with bootstrap result and CIs
\item{term}{variable name}
\item{mean_estimate}{mean of regression coefficients across bootstrap samples}
\item{conf.low}{lower 2.5 percentile bootstrap interval}
\item{conf.high}{upper 97.5 percentile bootstrap interval}
\item{median_p.value}{median p value of regression coefficients  across bootstrap samples}
\item{ci_ln}{confidence interval length}
\item{prop.select}{propotion of times a given variable is selected by model selection method}
\item{prop.rej}{proportion of time coefficient was found significant at 0.05 alpha level}
}
\description{
Inference for \code{selector} via bootstrapping

Bootstrapping for selection process with stepwise AIC/BIC
}
