% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_internal.R
\name{boot_pen}
\alias{boot_pen}
\title{Internal function for bootstrapping penalized model}
\usage{
boot_pen(
  x,
  y,
  B,
  family,
  std,
  nonselection,
  lambda_selected,
  lmax,
  all_terms,
  penalty,
  alpha,
  selected_terms,
  lambda_seq,
  nlambda,
  boot_desparse,
  save_beta
)
}
\description{
Internal function for bootstrapping penalized model
}
