% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_internal.R
\name{boot_stepwise_parallel}
\alias{boot_stepwise_parallel}
\title{Internal function for bootstrapping}
\usage{
boot_stepwise_parallel(
  x,
  y,
  penalty,
  B,
  family,
  n_cores,
  nonselector,
  direction,
  make_levels,
  save_beta,
  ...
)
}
\description{
Internal function for bootstrapping
}
