% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collat_fn.R
\name{ciratio}
\alias{ciratio}
\title{Computing mean/median CI ratio and significant Discoveries}
\usage{
ciratio(
  x,
  y,
  nonselection = "ignored",
  std = TRUE,
  B = 250,
  direction = "forward"
)
}
\arguments{
\item{x}{Dataframe/model matrix with predictors (without intercept)}

\item{y}{outcome vector}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{std}{if TRUE (default), standardize design matrix}

\item{B}{#of bootstraps}

\item{direction}{the mode of step wise search, can be one of "both", "backward", or "forward", with a default of "forward"}
}
\value{
A list with following data frames
\item{avg_ci_ln}{Mean of CI length across all variables in model excluding intercept}
\item{med_ci_ln}{Median of CI length across all variables in model excluding intercept}
\item{no_sign_disc}{# of significant discoveries in model excluding intercept}
}
\description{
Perfroms full model boostrap, stepwise AIC, stepwise BIc, lasso and MCP and gives average CI for variables
adjusting for post-selectino inference using hybrid, bootstrap and selective inference
}
