% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertain_nulls.R
\name{get_uncertain_nulls}
\alias{get_uncertain_nulls}
\title{Title}
\usage{
get_uncertain_nulls(mod, res, x)
}
\arguments{
\item{mod}{model results in data frame with NA for estimates for variables that are
not seelcted}

\item{res}{residual from final selected model}

\item{x}{Design data frame}
}
\value{
tidy data frame with coefficients for all variables
}
\description{
Title
}
