% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_inferrer.R, R/infer.R,
%   R/infer_pipe.R, R/infer_selective.R, R/infer_upsi.R, R/select_twostep_pval.R
\name{as_inferrer}
\alias{as_inferrer}
\alias{infer}
\alias{infer_pipe}
\alias{infer_selective}
\alias{infer_upsi}
\alias{infer.selector_two_step_pval}
\title{Constructor for \code{inferrer}s}
\usage{
as_inferrer(
  x,
  name,
  label = name,
  nonselection,
  inferences,
  conf.level,
  selector,
  meta = list()
)

infer(
  object,
  data,
  nonselection = c("ignored", "uncertain_nulls", "confident_nulls"),
  ...
)

infer_pipe(
  model,
  method = "hybrid",
  nonselection = "ignored",
  B = 250,
  n_cores = 1,
  save_beta = FALSE,
  boot_desparse = FALSE
)

infer_selective(
  model,
  method = "hybrid",
  nonselection = "ignored",
  B = 250,
  n_cores = 1,
  save_beta = FALSE,
  boot_desparse = FALSE
)

infer_upsi(
  object,
  data,
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls"),
  conf.level = 0.95
)

\method{infer}{selector_two_step_pval}(
  model,
  method = c("hybrid", "selectiveinf", "boot"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls")
)
}
\arguments{
\item{x}{a slot for the main inferential object}

\item{name}{brief name for method}

\item{label}{label for method}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{inferences}{a slot for post-selection inferences that should
have confidence intervals (at least)}

\item{conf.level}{.95 by default}

\item{selector}{a carried-forward \code{selector} object}

\item{meta}{a list containing important meta-information from the method}

\item{object}{a \code{selector} object}

\item{data}{data must be passed to \code{infer} methods}

\item{...}{additional arguments passed to downstream \verb{infer_*}}

\item{model}{model returned from selector_ic  class}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid"or  "selectiveinf"}
}
\value{
An S3 object (list) of class \code{inferrer}

A list of class \verb{infer_*} containing:'
\item{model}{A data frame with post-selection inference results}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{nonselection}{method chosen  to deal with non selection}
\item{infmethod}{Inference method chosen}
\item{selection_method}{Stepwsie,returned for selector_stepwise_ic class only}
\item{direction}{the mode of step wise search, returned for selector_stepwise_ic class only}
\item{penalty}{penalty used (AIC or BIC), returned for selector_stepwise_ic class only}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"; returned for selector_pen class only}
\item{alpha}{selected alpha for inference, returned for selector_pen class only}
\item{B}{The number of bootstrap replicates used (only for bootstrap selection method)}

"infer_pen_cv " class list with

"infer_pen_cv " class list with

\code{inferrer} object

infer_ic class list with
\item{model}{A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{medain CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
}
\description{
Create an \code{inferrer} wrapper around post-selection inference algorithms so they
can be used with the \code{selectInferToolkit} package and can harmonize results
across various selection algorithms (\code{inferrer}s). This is not a user-facing
function.

This function performs a \code{selector}'s default inference method. Alternative
inference methods may be available via \verb{infer_*}, which you should use
instead if you want to do specific tuning.

UPSI is sometimes referred to as "hybrid-OLS", but essentially we
re-fit the selected model to the data as though we never used that
same data to fit the model. It is common, easy, and ill-advised.
}
