% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.selector_ic}
\alias{infer.selector_ic}
\alias{infer.selector_pen}
\alias{infer.boot_selector_ic}
\alias{infer.boot_selector_pen}
\alias{infer}
\title{Inference for selector_ic class except bootstrap which is it's own function}
\usage{
\method{infer}{selector_ic}(
  model,
  method = c("hybrid", "selectiveinf", "boot"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls")
)

\method{infer}{selector_pen}(
  model,
  method = "hybrid",
  nonselection = "ignored",
  B = 250,
  n_cores = 1,
  save_beta = FALSE,
  boot_desparse = FALSE
)

\method{infer}{boot_selector_ic}(
  model,
  B = 250,
  nonselection = c("ignored", "uncertain", "confident_nulls"),
  n_cores = 1,
  make_levels = FALSE,
  save_beta = FALSE,
  ...
)

\method{infer}{boot_selector_pen}(model, B, nonselection, n_cores = 1, save_beta = FALSE, ...)

infer(
  model,
  method = c("hybrid", "boot", "selectiveinf", "PIPE"),
  nonselection = c("ignored", "uncertain_nulls", "confident_nulls"),
  ...
)
}
\arguments{
\item{model}{model of selector_ic or selector_pen class returned from step_ic or pen_cv function}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid", "selectiveinf" or
"boot" supported}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}

\item{B}{The number of bootstrap replicates.}

\item{n_cores}{whether to run bootstrap in parallel for faster computation}

\item{...}{}
}
\value{
infer_ic class list with
\item{model}{A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{medain CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}

"infer_pen" class list with

"boot_ic" class list with
\item{model}{A dataframe with model results including columns for term, mean_estimate,  conf.low, conf.high, median_p.value,ci_ln, prop.select, prop.reject}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{selected the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
\item{B}{The number of bootstrap replicates used}

"infer_boot_selector_pen" class list with
\item{model}{A dataframe with model results including columns for term, mean_estimate,  conf.low, conf.high,ci_ln, prop.select}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{selection_method}{lasso' or 'MCP' Note: Same as in ncvreg, even with elasticnet-it says lasso but can be differntiated with value of alpha}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"}
\item{alpha}{selected alpha for inference}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
\item{B}{The number of bootstrap replicates used}

A list of class \code{infer_ic} or \code{infer_pen} containing:#'
\item{model}{A data frame with post-selection inference results}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{nonselection}{method chosen  to deal with non selection}
\item{infmethod}{Inference method chosen}
\item{selection_method}{Stepwsie,returned for selector_ic class only}
\item{direction}{the mode of step wise search, returned for selector_ic class only}
\item{penalty}{penalty used (AIC or BIC), returned for selector_ic class only}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"; returned for selector_pen class only}
\item{alpha}{selected alpha for inference, returned for selector_pen class only}
\item{B}{The number of bootstrap replicates used (only for bootstrap selection method)}
}
\description{
Inference for selector_ic class except bootstrap which is it's own function

Inference for selector_pen class except bootstrap which is it's own function

Inference for selector_ic class based on bootstrap
other methods are supposed in infer.selector_ic

Title Inference for selector_pen class based on bootstrap, other methods are supposed in infer.selector_pen

Post-selection inference
}
