% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Post-selection inference}
\usage{
infer(model, method = "hybrid", nonselection = "ignored", ...)
}
\arguments{
\item{model}{model of selector_ic or selector_pen class returned from step_ic or pen_cv function}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid", "selectiveinf" or
"boot" supported}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
A list of class \code{infer_ic} or \code{infer_pen} containing:#'
\item{model}{A data frame with post-selection inference results}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{nonselection}{method chosen  to deal with non selection}
\item{infmethod}{Inference method chosen}
\item{selection_method}{Stepwsie,returned for selector_ic class only}
\item{direction}{the mode of step wise search, returned for selector_ic class only}
\item{penalty}{penalty used (AIC or BIC), returned for selector_ic class only}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"; returned for selector_pen class only}
\item{alpha}{selected alpha for inference, returned for selector_pen class only}
\item{B}{The number of bootstrap replicates used (only for bootstrap selection method)}
}
\description{
Post-selection inference
}
