% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R, R/pen_cv.R, R/stepwise_ic.R,
%   R/two_step_pval.R
\name{infer}
\alias{infer}
\alias{infer.selector_pen}
\alias{infer.selector_stepwise_ic}
\alias{infer.selector_two_step_pval}
\title{Post-selection inference}
\usage{
infer(
  model,
  method = c("hybrid", "boot", "selectiveinf", "PIPE"),
  nonselection = c("ignored", "uncertain_nulls", "confident_nulls"),
  ...
)

\method{infer}{selector_pen}(
  model,
  method = "hybrid",
  nonselection = "ignored",
  B = 250,
  n_cores = 1,
  save_beta = FALSE,
  boot_desparse = FALSE
)

\method{infer}{selector_stepwise_ic}(
  model,
  method = c("hybrid", "selectiveinf", "boot"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls"),
  B = 250,
  n_cores = 1,
  save_beta = FALSE,
  ...
)

\method{infer}{selector_two_step_pval}(
  model,
  method = c("hybrid", "selectiveinf", "boot"),
  nonselection = c("ignored", "confident_nulls", "uncertain_nulls")
)
}
\arguments{
\item{model}{model returned from selector_ic  class}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid"or  "selectiveinf"}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
A list of class \verb{infer_*} containing:'
\item{model}{A data frame with post-selection inference results}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{nonselection}{method chosen  to deal with non selection}
\item{infmethod}{Inference method chosen}
\item{selection_method}{Stepwsie,returned for selector_stepwise_ic class only}
\item{direction}{the mode of step wise search, returned for selector_stepwise_ic class only}
\item{penalty}{penalty used (AIC or BIC), returned for selector_stepwise_ic class only}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"; returned for selector_pen class only}
\item{alpha}{selected alpha for inference, returned for selector_pen class only}
\item{B}{The number of bootstrap replicates used (only for bootstrap selection method)}

"infer_pen" class list with

infer_stepwise_ic class list with
\item{model}{A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{medain CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}

infer_ic class list with
\item{model}{A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{medain CI length across all variables in model}
\item{selection_method}{Stepwsie}
\item{direction}{the mode of step wise search}
\item{penalty}{penalty used (AIC or BIC)}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
}
\description{
Post-selection inference

Inference for selector_pen class except bootstrap which is it's own function

Inference for selector_stepwise_ic class except bootstrap which is it's own function

Inference for selector_ic class except bootstrap which is it's own function
}
