% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer.selector.pen}
\alias{infer.selector.pen}
\title{Inference for selector.pen class except bootstrap which is it's own function}
\usage{
infer.selector.pen(model, method = "hybrid", nonselection = "ignored")
}
\arguments{
\item{model}{model of selector_pen class returned from  pen_cv function}

\item{method}{A character string specifying method of post-selection inference.Currently "hybrid", "selectiveinf" or
boot supported}

\item{nonselection}{A character string specifying how to handle variables not selected by model selection procedure. One of
"ignored", "confident_nulls" or "uncertain_nulls" supported}
}
\value{
"infer_pen" class list with
\item{model}{A dataframe with model results including columns for term, estimate, std.error, statistics, p.value, conf.low, conf.high, ci_ln}
\item{ci_avg_ratio}{Average CI length across all variables in model}
\item{ci_median_ratio}{median CI length across all variables in model}
\item{selection_method}{'lasso' or 'MCP' Note: Same as in ncvreg, even with elastic net-it says lasso but can be differntiated with value of alpha}
\item{lambda}{selected lambda for inference , either "lambda.min" or "lambda.1se"}
\item{alpha}{selected alpha for inference}
\item{infmethod}{Inference method chosen}
\item{nonselection}{method chosen  to deal with non selection}
}
\description{
Inference for selector.pen class except bootstrap which is it's own function
}
