% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_cv.R
\name{pen_cv}
\alias{pen_cv}
\title{Fit an MCP- or SCAD-penalized regression}
\usage{
pen_cv(
  x,
  y,
  family = "gaussian",
  std = TRUE,
  penalty = "lasso",
  lambda = "lambda.min",
  alpha = 1,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{x}{Dataframe/model matrix with predictors (without intercept)}

\item{y}{outcome vector}

\item{family}{currently gaussian supported}

\item{std}{if TRUE (default), standardize design matrix}

\item{penalty}{lasso or MCP}

\item{lambda}{extra coefficients associated with "lambda.min" or "lambda.1se"}

\item{alpha}{Tuning parameter  small, but not exactly 0.}

\item{...}{Additional arguments that can be passed with cv.ncvreg function in cv.ncvreg package}
}
\value{
A list of class \code{selector_pen} containing:
\item{beta}{a tibble containing term names and coefficients}
\item{std}{Was desing matrix standadrized}
\item{penalty}{penalty used (lasso or MCP)}
\item{lambda}{are the coefficeint associated with "lambda.min" or "lambda.1se"}
\item{lambda.select}{numeric value of selected lambda}
\item{fold}{Which fold each observation belongs to. By default the observations are randomly assigned.}
\item{x}{ the model dataframe used}
\item{y}{repsonse used}
\item{alpha}{selected alpha for model fitting}
}
\description{
Performs k-fold cross validation for MCP- or SCAD-penalized regression
models over a grid of values for the regularization parameter lambda and returns the
coefficeints associated with either value of lambda that gives minimum cvm or largest value of lambda such that error is within 1 standard error of the minimum.
}
