% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selec_inf.R
\name{sel_inf}
\alias{sel_inf}
\title{Title}
\usage{
sel_inf(x, y, lam = "lambda.min", std = FALSE, intercept = TRUE, ...)
}
\arguments{
\item{x}{design matrix}

\item{y}{outcome variable}

\item{lam}{lambda.1se or lambda.min}

\item{std}{whether to standaddize design matrix}

\item{...}{addtional arguments that can be passed to  cv.glmnet}
}
\value{
returns a data frame with results from lasso selective inference
}
\description{
Title
}
