% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selec_inf.R
\name{sel_inf_fs}
\alias{sel_inf_fs}
\title{Title}
\usage{
sel_inf_fs(x, y, mult = 2, intercept = TRUE, std = F, ...)
}
\arguments{
\item{x}{design matrix}

\item{y}{outcome variable}

\item{mult}{}
}
\value{
returns a data frame with results from forward stepwise selective inference
}
\description{
Title
}
