% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_glmnet.R
\name{select_glmnet}
\alias{select_glmnet}
\title{Fit an (relaxed) lasso or elastic net penalized regression via glmnet}
\usage{
select_glmnet(
  formula,
  data,
  family = c("gaussian", "binomial", "poisson"),
  lambda = c("best", "compact"),
  fitted_selector = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula}

\item{data}{data set}

\item{family}{outcome distributional family}

\item{lambda}{can be \code{best}, \code{compact} (which use CV), or a numeric vector. See details.}

\item{fitted_selector}{a previously fit \code{selector}, used for resampling}

\item{...}{Additional arguments that can be passed to \code{glmnet}, e.g. \code{alpha}}
}
\value{
A \code{selector} object wrapping \code{glmnet} containing:
\item{beta}{a tibble containing term names and coefficients}
\item{std}{Was desing matrix standadrized}
\item{penalty}{penalty used (lasso or MCP)}
\item{lambda}{are the coefficeint associated with "lambda.min" or "lambda.1se"}
\item{lambda.select}{numeric value of selected lambda}
\item{fold}{Which fold each observation belongs to. By default the observations are randomly assigned.}
\item{x}{ the model dataframe used}
\item{y}{repsonse used}
\item{alpha}{selected alpha for model fitting}
}
\description{
A wrapper for \code{glmnet} and \code{cv.glmnet}. It runs CV by default so remember to
set your seed for reproducibility.
}
