% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_stepwise_ic.R
\name{select_stepwise_ic}
\alias{select_stepwise_ic}
\title{Stepwise forward/backward/bidirectional selection with AIC/BIC}
\usage{
select_stepwise_ic(
  formula,
  data,
  family = c("gaussian", "binomial"),
  select_factors_together = TRUE,
  penalty = c("AIC", "BIC"),
  direction = c("forward", "backward", "both"),
  trace = 0,
  fitted_selector = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula.}

\item{data}{data set}

\item{family}{outcome distributional family}

\item{select_factors_together}{should categorical variables be jointly selected?}

\item{penalty}{AIC, BIC}

\item{direction}{the mode of step wise search, can be one of "both", "backward", or "forward", with a default of "forward"}

\item{trace}{passed to MASS::stepAIC}

\item{fitted_selector}{a previously fit \code{selector}, used for resampling}

\item{...}{Additional arguments}
}
\value{
A  \code{selector} object
}
\description{
This function implements forward/backward/bidirectional stepwise regression,
for use in the \code{selectInferToolkit} package
}
