% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector.R
\name{as_selector}
\alias{as_selector}
\alias{predict.selector}
\alias{tidy.selector}
\alias{print.selector}
\alias{coef.selector}
\alias{reselect}
\title{Constructor for \code{selector}s}
\usage{
as_selector(
  x,
  name,
  label = name,
  all_terms,
  recipe_obj,
  selected_coefs,
  default_infer,
  meta = list()
)

\method{predict}{selector}(object, newdata = NULL, ...)

\method{tidy}{selector}(x, ...)

\method{print}{selector}(x, ...)

\method{coef}{selector}(x)

reselect(selector_obj, newdata)
}
\arguments{
\item{x}{an object from a selection algorithm}

\item{all_terms}{a slot containing names of all terms}

\item{selected_coefs}{a numeric vector of selected coefficient estimates}

\item{meta}{a list containing important meta-information}
}
\value{
An S3 object (list) of class \code{selector} containing:'
x
all_terms
meta
}
\description{
Create a \code{selector} wrapper around selection algorithms so they
can be used with the \code{selectInferToolkit} package.

Returns predictions for \code{selector} objects on new data

Returns a tibble with all candidate variables, estimates (scaled & unscaled)

Returns a vector with all selected coefficients (scaled only, includes intercept)

Re-do the selection, possibly on new data. Does not re-do the pre-processing;
instead uses the original preprocess recipe from \code{selector_obj}.
}
