% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ic.R
\name{step_ic}
\alias{step_ic}
\title{Stepwise forwad/backward/bidirectional selection with AIC/BIC}
\usage{
step_ic(
  x,
  y,
  family = "gaussian",
  std = FALSE,
  penalty = "AIC",
  direction = "forward",
  make_levels = FALSE,
  ...
)
}
\arguments{
\item{x}{Dataframe/model matrix with predictors (without intercept)}

\item{y}{outcome vector}

\item{family}{currently gaussian supported}

\item{std}{if TRUE (default), standardize design matrix}

\item{penalty}{AIC or BIC}

\item{direction}{the mode of step wise search, can be one of "both", "backward", or "forward", with a default of "forward"}

\item{make_levels}{whether to model selection after dummy coding for categorical variables (defult FALSE)}

\item{...}{Additional arguments that can be passed with stepAIC function in MASS package}
}
\value{
A list of class \code{selector_ic} containing:#'
\item{beta}{a tibble containing term names and coefficients}
\item{std}{Was desing matrix standadrized}
\item{penalty}{penalty used (AIC or BIC)}
\item{direction}{the mode of step wise search}
\item{x}{the model dataframe used}
\item{y}{repsonse used in vector}
\item{model_sum}{the stepwise-selected model  details is returned}
}
\description{
This function implements forward/backward/bidirectional stepwise regression,
for use in the selectInferToolkit package
}
